more int type for init changes

+++ linbox/algorithms/lifting-container.h
@@ -365,7 +365,7 @@ namespace LinBox
 			typename Vector1::const_iterator     b_iter    = b.begin();
 			typename BlasVector<Ring>::iterator  res_iter  = _b.begin() ;
 			for (; b_iter != b.end(); ++res_iter, ++b_iter)
-				this->_intRing.init(*res_iter, *b_iter);
+				this->_intRing.init(*res_iter, int64_t(*b_iter));
 
 			Integer_t had_sq, short_sq;
 			BoundBlackbox(this->_intRing, had_sq, short_sq, A);
